WalletApplication/src/main/java/com/capgemini/walletapplication/bean/AccountDetails.java
package com.capgemini.walletapplication.bean;

import java.time.LocalDate;
import java.util.ArrayList;
import java.util.HashMap;
import java.util.List;
import java.util.Map;

public class AccountDetails {

	
	private long accNo;
	LocalDate date;
	private String username;
	private String password;
	private double balance;
	private long transId;
	private String branch;
	private String accType;
	
	private Customer customer;
	
	public long getAccNo() {
		return accNo;
	}
	public void setAccNo(long accNo) {
		this.accNo = accNo;
	}

	public LocalDate getDate() {
		return date;
	}
	public void setDate(LocalDate date) {
		this.date = date;
	}
	
	public String getUsername() {
		return username;
	}
	public void setUsername(String username) {
		this.username = username;
	}
	public String getPassword() {
		return password;
	}
	public void setPassword(String password) {
		this.password = password;
	}
	
	
	public double getBalance() {
		return balance;
	}


	public void setBalance(double balance) {
		this.balance = balance;
	}
	public long getTransId() {
		return transId;
	}

	public void setTransId(long transId) {
		this.transId = transId;
		
	}
	public Customer getCustomer() {
		return customer;
	}
	public void setCustomer(Customer customer) {
		this.customer = customer;
	}
	
	
}
